move wrapper into internal
