+++ instrumentation/github.com/gorilla/mux/otelmux/internal/request/body_wrapper.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package request // import "go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/request"
+package request // import "go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/request"
 
 import (
 	"io"
