+++ instrumentation/github.com/gorilla/mux/otelmux/mux.go
@@ -11,7 +11,7 @@ import (
 	"github.com/gorilla/mux"
 
 	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/semconvutil"
-	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/request"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gorilla/mux/otelmux/internal/request"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/propagation"
 	semconv "go.opentelemetry.io/otel/semconv/v1.20.0"
